backend/requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain
beautifulsoup4
requests
pydantic
python-dotenv
